// src/utils/types.ts
import { ReactNode } from "react";

export type CartItem = {
  id: number;
  quantity: number;
};
export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: {
    rate: number;
    count: number;
  };
}

export type CartProviderProps = {
  children: ReactNode;
};

export type CartContext = {
  getItemQuantity: (id: number) => number;
  increaseCartQuantity: (id: number) => void;
  decreaseCartQuantity: (id: number) => void;
  removeFromCart: (id: number) => void;
  cartQuantity: number;
  cartItems: CartItem[];
  products: Product[];
  clearCart: () => void;
};